+++ spec/rails_helper.rb
@@ -40,7 +40,7 @@ RSpec.configure do |config|
   # If you're not using ActiveRecord, or you'd prefer not to run each of your
   # examples within a transaction, remove the following line or assign false
   # instead of true.
-  # config.use_transactional_fixtures = true
+  config.use_transactional_fixtures = false
 
   # RSpec Rails can automatically mix in different behaviours to your tests
   # based on their file location, for example enabling you to call `get` and
@@ -59,22 +59,4 @@ RSpec.configure do |config|
 
   config.include Devise::TestHelpers,           type: :controller
   config.include IntegrationHelpers,            type: :feature
-
-  # Turn this off in all request specs
-  module DisableTransactionalFixtures
-    def self.included(base)
-      base.use_transactional_fixtures = false
-      base.after(:each) do |example|
-        while true
-          begin
-            DatabaseCleaner.clean
-            break
-          rescue Exception => e
-            sleep 2
-          end
-        end
-      end
-    end
-  end
-  config.include DisableTransactionalFixtures,  type: :feature
 end