+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - Remove the API module used in `using` params into Project Entity
+- Clean up DatabaseCleaner config
 
 ## [1.3.0] - 2017-04-25
 ### Added